# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tarn AI Evaluation API",
    "description": "AI suggests; the teacher decides. Every college sees only its own data.",
    "version": "0.0.0"
  },
  "paths": {
...
    "/api/v1/booklets/{booklet_id}/segments": {
      "get": {
        "tags": [
          "segments"
        ],
        "summary": "A booklet's segments in written order (unassigned ones have no question)",
        "operationId": "list_segments_api_v1_booklets__booklet_id__segments_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentsOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/segments/merge": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "Merge two segments (the second joins the first)",
        "operationId": "merge_api_v1_booklets__booklet_id__segments_merge_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentsOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/segments/split": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "Split a segment: the regions from `at_region` on become a new segment",
        "operationId": "split_api_v1_booklets__booklet_id__segments_split_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SplitIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentsOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/segments/reassign": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "Give a segment another question, or move it to the unassigned tray",
        "operationId": "reassign_api_v1_booklets__booklet_id__segments_reassign_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReassignIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentsOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/booklets/{booklet_id}/segments/move-boundary": {
      "post": {
        "tags": [
          "segments"
        ],
        "summary": "Move the boundary between two neighbouring segments",
        "operationId": "move_boundary_api_v1_booklets__booklet_id__segments_move_boundary_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "booklet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Booklet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MoveBoundaryIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegmentsOut"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "description": "Stale version (reload), a move the workflow does not allow, a new suggestion still on its way, or an approved answer that must be reopened first.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "422": {
            "description": "A mark, tag or text the rules refuse.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorOut"
                }
              }
            }
          },
          "423": {
            "description": "Another teacher has the booklet open, or the caller has not opened it.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LockedOut"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "ErrorOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LockedOut": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "holder_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Holder Id",
            "description": "Who has the booklet open (GET .../review names them)."
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "holder_id",
          "expires_at"
        ],
        "title": "LockedOut"
      },
...
      "MergeIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "first": {
            "type": "string",
            "format": "uuid",
            "title": "First"
          },
          "second": {
            "type": "string",
            "format": "uuid",
            "title": "Second"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "first",
          "second"
        ],
        "title": "MergeIn"
      },
      "MoveBoundaryIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "upper": {
            "type": "string",
            "format": "uuid",
            "title": "Upper"
          },
          "lower": {
            "type": "string",
            "format": "uuid",
            "title": "Lower"
          },
          "region": {
            "type": "string",
            "format": "uuid",
            "title": "Region",
            "description": "The first region of `lower` after the move."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "upper",
          "lower",
          "region"
        ],
        "title": "MoveBoundaryIn"
      },
...
      "ReassignIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "segment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Segment Id"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "None: the unassigned tray."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "segment_id"
        ],
        "title": "ReassignIn"
      },
...
      "SegmentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "slot_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slot Label",
            "description": "None: the unassigned tray."
          },
          "proposed_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Label"
          },
          "position": {
            "type": "integer",
            "title": "Position"
          },
          "source": {
            "type": "string",
            "enum": [
              "rule",
              "similarity",
              "teacher"
            ],
            "title": "Source"
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags"
          },
          "match_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Match Score"
          },
          "region_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Region Ids"
          },
          "page_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Page Ids"
          }
        },
        "type": "object",
        "required": [
          "id",
          "slot_label",
          "proposed_label",
          "position",
          "source",
          "flags",
          "match_score",
          "region_ids",
          "page_ids"
        ],
        "title": "SegmentOut"
      },
      "SegmentsOut": {
        "properties": {
          "booklet_version": {
            "type": "integer",
            "title": "Booklet Version"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/SegmentOut"
            },
            "type": "array",
            "title": "Segments"
          },
          "rescoring": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Rescoring"
          },
          "emptied": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Emptied",
            "description": "Answers left without text (not attempted)."
          }
        },
        "type": "object",
        "required": [
          "booklet_version",
          "segments"
        ],
        "title": "SegmentsOut"
      },
...
      "SplitIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Expected Version"
          },
          "segment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Segment Id"
          },
          "at_region": {
            "type": "string",
            "format": "uuid",
            "title": "At Region"
          },
          "label": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "None: the unassigned tray."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "segment_id",
          "at_region"
        ],
        "title": "SplitIn"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Access token from /api/v1/auth/login",
        "scheme": "bearer"
      }
    }
  }
}